[Feat] 실습 페이지 분리
아코디언 형식으로 접기/펴기 기능 추가

+++ src/hooks/useButtonHandler.tsx
@@ -2,33 +2,63 @@ import useUserStore from '../store/userStore';
 import { createStack, deleteStack, getConsoleUrl } from '../utils/api';
 import { useNavigate } from 'react-router-dom';
 import useStackInfoPoller from './useStackInfoPoller';
+import { toast } from 'react-toastify';
+import useStackStore from '../store/stackStore';
 
-const useButtonHandler = (practiceId: string | undefined) => {
+const useButtonHandler = () => {
   const navigate = useNavigate();
   const { userId } = useUserStore();
-  const { startPolling } = useStackInfoPoller(practiceId);
-  // const { setStack } = useStackStore();
+  const { practiceId } = useStackStore();
+  const { startPolling } = useStackInfoPoller();
+  
+  // 람다 깨우는 시간 때문에 await 응답이 늦어 요청 toast를 띄움. 추후 상의필요 공민혁
+  /**
+   * 스택 생성 버튼 클릭 동작
+   * 스택 생성 api 요청 후 스택 상태 확인
+   */
+  const handleStackCreation = async () => {
+    try {
+      if (window.confirm('실습 환경을 생성하시겠습니까?')) {
+        toast.success('실습 환경 생성 요청');
+        await createStack(practiceId, userId);
+        toast.success('실습 환경 생성 중입니다!');
+      }
 
-  const handleStackCreation = async (practiceId: string | undefined) => {
-    await createStack(practiceId, userId);
-
-    // api 호출 성공 시 임의로 스택 상태값 바꿔 useStackInfoPoller 트리거
-    startPolling();
-    // setStack('status', StackStatusSchema.Values.CREATE_IN_PROGRESS);
+      startPolling();
+    } catch {
+      toast.error('실습 환경 생성에 실패했습니다.');
+    }
   };
 
-  const handleOpenConsole = async (practiceId: string | undefined) => {
-    const url = await getConsoleUrl(practiceId, userId);
-    window.open(url, '_blank');
+  /**
+   * 콘솔 접근 버튼 클릭 동작
+   * url api 요청 후 해당 url 새 창에서 오픈
+   */
+  const handleOpenConsole = async () => {
+    try {
+      const url = await getConsoleUrl(practiceId, userId);
+      window.open(url, '_blank');
+    } catch {
+      toast.error('콘솔 접근에 실패했습니다.');
+    }
   };
 
-  const handleStackDeletion = async (practiceId: string | undefined) => {
-    await deleteStack(practiceId, userId);
-
-    // api 호출 성공 시 임의로 스택 상태값 바꿔 useStackInfoPoller 트리거
-    startPolling();
-    // setStack('status', StackStatusSchema.Values.DELETE_IN_PROGRESS);
-    navigate('/');
+  // 람다 깨우는 시간 때문에 await 응답이 늦어 요청 toast를 띄움. 추후 상의필요 공민혁
+  /**
+   * 스택 삭제 버튼 클릭 동작
+   * 스택 삭제 api 요청 후 메인 페이지로 이동
+   */
+  const handleStackDeletion = async () => {
+    try {
+      if (window.confirm('실습 환경을 삭제하시겠습니까?')) {
+        toast.success('실습 환경 삭제 요청');
+        await deleteStack(practiceId, userId);
+        toast.success('실습 환경 삭제 중입니다.');
+        navigate('/');
+      }
+    } catch {
+      toast.error('실습 환경 삭제에 실패했습니다.');
+    }
   };
 
   return { handleStackCreation, handleOpenConsole, handleStackDeletion };
